+++ library/Episciences/Review.php
@@ -355,6 +355,7 @@ class Episciences_Review
 
             if (!$role) {
                 self::checkReviewNotifications($cc);
+                Episciences_Submit::addIfNotExists($paper->getEditors(), $cc);
                 Episciences_PapersManager::keepOnlyUsersWithoutConflict($paper->getPaperid(), $cc);
 
             } elseif ($role === Episciences_Acl::ROLE_REVIEWER) {