+++ library/Episciences/Paper.php
@@ -5196,6 +5196,10 @@ class Episciences_Paper
 
     public function getMainPaperUrl(): ?string
     {
+        if($this->isTmp()){
+            return null;
+        }
+
         if ($this->isDataSetOrSoftware()) {
             return $this->getDataDescriptorUrl();
         }