+++ application/modules/common/controllers/DefaultController.php
@@ -1,6 +1,7 @@
 <?php
 use GuzzleHttp\Client;
 use GuzzleHttp\Exception\GuzzleException;
+use Monolog\Logger;
 
 
 class DefaultController extends Zend_Controller_Action
@@ -382,5 +383,4 @@ class DefaultController extends Zend_Controller_Action
         $this->view->error = true;
 
     }
-
 }