+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - published dataset/software: the title of the data/software descriptor is its identifier.
 - Only PDF files are backed up
 ### Fixed
+- [#653](https://github.com/CCSDForge/episciences/issues/653): Filtering email history doesn't work
 - File Not Found Error
 ## v1.0.50.1 - 2025-03-14
 ### Fixed