+++ CHANGELOG.md
@@ -15,7 +15,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Security
 -->
 ## Unreleased
-
+ 
 ### Changed
 - [#560](https://github.com/CCSDForge/episciences/issues/650): send notifications when a cover letter is added or edited