+++ CHANGELOG.md
@@ -16,7 +16,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 -->
 ## Unreleased
 
+
 ### Changed
+- [#650] pas de notifications quand un commentaire est ajouté par l'auteur
 - [#560](https://github.com/CCSDForge/episciences/issues/650): send notifications when a cover letter is added or edited
 
 - Ability to customise application environment variables
@@ -30,8 +32,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Only PDF files are backed up
 - [#649](https://github.com/CCSDForge/episciences/issues/649): Assigned editors are now automatically added as (hidden) copies of messages sent to reviewers
 ### Fixed
-- reset password url
-- account creation token url
+- reset password url (Manager only)
+- account creation token url (Manager only)
 - [#653](https://github.com/CCSDForge/episciences/issues/653): Filtering email history doesn't work
 - File Not Found Error
 - Wrong link on page "Invitation to review an article" for temporary versions