+++ CHANGELOG.md
@@ -14,18 +14,20 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Removed
 ### Security
 -->
-## Unreleased
-- [#560](https://github.com/CCSDForge/episciences/issues/650): send notifications when a cover letter is added or edited
-- [#650] pas de notifications quand un commentaire est ajouté par l'auteur 
 
+v1.0.51 - 2025-04-07
 ### Changed
+- [#650](https://github.com/CCSDForge/episciences/issues/650): send notifications when a cover letter is added or edited
 - published dataset/software: the title of the data/software descriptor is its identifier.
 - Only PDF files are backed up
 - [#649](https://github.com/CCSDForge/episciences/issues/649): Assigned editors are now automatically added as (hidden) copies of messages sent to reviewers
+- A copy of the default grid is proposed if the default grid is not empty, and if the destination grid is not empty either.
+
 ### Fixed
 - [#653](https://github.com/CCSDForge/episciences/issues/653): Filtering email history doesn't work
 - File Not Found Error
 - Wrong link on page "Invitation to review an article" for temporary versions
+
 ## v1.0.50.1 - 2025-03-14
 ### Fixed
 - Fixed CACHE_PATH not existing